It would be enough
Do not you agree? OK. Imagine two worlds: in one there is no any commandment at all and in the other The Only commandment is This One. In which world you would like to live? 
Nope? Do you still disagree? Do you have a better One? OK. Imagine two worlds: in one of them The Only commandment is yours and in the other - The Only One is Mine. Which world you would prefer to live in? 
No rush, take your time...